Factor out TranslationDiff component
