+++ frontend/src/core/diff/index.js
@@ -1,0 +1,3 @@
+/* @flow */
+
+export { default as TranslationDiff } from './components/TranslationDiff';
